Refactor: Split IR into files
