+++ src/IR/index.ts
@@ -1,3 +1,2 @@
-export * from "./builders";
 export * as IR from "./IR";
-export * from "./traverse";
+export * from "./IR";
