+++ src/cli.ts
@@ -12,7 +12,7 @@ import {
   program,
   whileLoop,
 } from "./IR/builders";
-import { programToPath } from "./IR/traverse";
+import { programToPath } from "./common/traverse";
 import lua from "./languages/lua";
 import debugEmit from "./languages/debug/emit";
 import { applyLanguage } from "./common/applyLanguage";
